requirements.txt
django==3.1
djangorestframework==3.11.0
psycopg2>=2.7
django-cors-headers
django-hashid-field